gateway/pyproject.toml
[project]
name = "hlsc-gateway"
version = "0.1.0"
description = "DataManager API 网关 — 转发并改写参数/返回值"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["routes"]
